refactor: replace error module with folder
