+++ aicompleter/error/__init__.py
@@ -1,0 +1,10 @@
+from .base import *
+from .special import *
+from .aigenerate import *
+
+__all__ = (
+    'BaseException',
+    *(
+        i.__class__.__name__ for i in globals().values() if issubclass(i, BaseException)
+    )
+)
